# api/swagger/swagger.yaml
swagger: "2.0"
...
info:
  version: "0.0.1"
  title: Education Consultant Assistant API

################################################################################
#                  Host, Base Path, Schemes and Content Types                  #
################################################################################
host: localhost:10010
...
basePath: /api/v0/
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json

################################################################################
#                                   Tags                                       #
################################################################################
...
parameters:
  workflowStatusId:
    name: workflowStatusId
    in: path
    description: The id of a workflow status
    type: string
    required: true
...
paths:
...
  /workflow-statuses:
    x-swagger-router-controller: workflowStatuses
    post:
      operationId: createWorkflowStatus
      tags:
        - Workflow Status
      description: ""
      summary: Creates a new workflow status type
      parameters:
        - name: newWorkflowStatus
          in: body
          description: The NewWorkflowStatus object
          required: true
          schema:
            $ref: "#/definitions/NewWorkflowStatus"
      responses:
        201:
          description: Created
          schema:
            $ref: "#/definitions/ExpandedWorkflowStatus"
        400:
          description: Bad request
          schema:
            $ref: "#/definitions/ErrorResponse"
        409:
          description: Conflict
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /workflow-statuses/{workflowStatusId}:
    x-swagger-router-controller: workflowStatuses
    parameters:
      - $ref: '#/parameters/workflowStatusId'
    get:
      operationId: getWorkflowStatusById
      tags:
        - Workflow Status
      description: ""
      summary: Returns a workflow status type
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/ExpandedWorkflowStatus"
        404:
          description: Not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      operationId: updateWorkflowStatus
      tags:
        - Workflow Status
      description: ""
      summary: Updates a workflow status type
      parameters:
        - name: name
          in: body
          description: The WorkflowStatus object
          required: true
          schema:
            $ref: "#/definitions/WorkflowStatus"
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/ExpandedWorkflowStatus"
        400:
          description: Bad request
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
    patch:
      operationId: partialUpdateWorkflowStatus
      tags:
        - Workflow Status
      description: ""
      summary: Partially update a workflow status
      consumes:
        - application/merge-patch+json
      parameters:
        - name: workflowStatusUpdatePatch
          in: body
          description: The WorkflowStatusUpdatePatch object
          required: true
          schema:
            $ref: "#/definitions/WorkflowStatusUpdatePatch"
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/ExpandedWorkflowStatus"
        400:
          description: Bad request
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId: deleteWorkflowStatus
      tags:
        - Workflow Status
      description: ""
      summary: Deletes a workflow status type
      responses:
        204:
          description: No content
        404:
          description: Not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"

  ##### Customer #####
...
definitions:
...
  NewWorkflowStatus:
    type: object
    required:
      - name
      - color
    properties:
      name:
        type: string
      color:
        $ref: "#/definitions/Color"
      pos:
        $ref: "#/definitions/SortablePosition"
    additionalProperties: false
  WorkflowStatus:
    type: object
    allOf:
      - $ref: "#/definitions/NewWorkflowStatus"
      - required:
        - _id
        - pos
        properties:
          _id:
            $ref: "#/definitions/ObjectId"
          pos:
            $ref: "#/definitions/SortablePosition"
  ExpandedWorkflowStatus:
    type: object
    allOf:
      - $ref: "#/definitions/WorkflowStatus"
      - required:
        - created_at
        - updated_at
        properties:
          created_at:
            type: string
            format: date-time
          updated_at:
            type: string
            format: date-time
  WorkflowStatusUpdatePatch:
    type: object
    properties:
      name:
        type: string
      color:
        $ref: "#/definitions/Color"
      pos:
        $ref: "#/definitions/SortablePosition"
    additionalProperties: false

  # Customer
...
  ObjectId:
    type: string
    pattern: ^[0-9a-fA-F]{24}$
  SortablePosition:
    type: integer
    format: int32
    minimum: 0

  # Enums
  Color:
    type: string
    enum:
      - red
      - orange
      - yellow
      - green
      - blue
      - purple
      - black
...
  ErrorResponse:
    type: object
    required:
      - statusCode
      - error
      - message
    properties:
      statusCode:
        type: integer
      error:
        type: string
      message:
        type: string
      details:
        type: array
        items:
          type: object